Create manus plugin
The manus plugin uses the Manus SDK to query glove data, and polls
the OpenXR for controller pose. It then uses an offset aim pose
as the wrist joint pose, and uses that as the root for all the
subsequent joint poses from Manus. Then it pushes this back to
the CloudXR runtime.