Create manus plugin (#14)
* Create manus plugin

The manus plugin uses the Manus SDK to query glove data, and polls
the OpenXR for controller pose. It then uses an offset aim pose
as the wrist joint pose, and uses that as the root for all the
subsequent joint poses from Manus. Then it pushes this back to
the CloudXR runtime.

* Fix improper xrtime handling

* Don't fail if manus SDK is not present

* Handle ManusSDK missing

* Remove obsolute comment

* Refactoring common code across plugins

* Refactoring plugin file structure

* Address MR feedback regarding dir layout and cmake files

* Create math.h utils, namespace the plugins

* Rewrite the manus plugin as it was copy-pasted from holoscan op

* Re-write the plugin utils for modern C++

* Addressing cmake issues

* Moving oxr ext header

* remove unused INSTALL_INTERFACE declariations

* Cleaning up main

* Removing unneeded check

* Addressing feedback